Move `DSLinear` to `layers.py`.
